feat: update check for updates logic
