+++ apps/mobile/src/services/index.ts
@@ -5,3 +5,4 @@ export * from './settings';
 export * from './watch-list';
 export * from './notifications';
 export * from './useAppLoadService';
+export * from './checkForUpdates';
